requirements.txt
fastapi
uvicorn[standard]
ai21
python-dotenv
pydantic